refactor: GameResult Component Barrel Pattern 적용 #439
