+++ frontend/src/pages/GameResultPage/components/GameResult/components/FinalButton/FinalButton.tsx
