+++ frontend/src/pages/GameResultPage/components/GameResult/components/GameResultItem/GameResultItem.styled.ts
