+++ frontend/src/pages/GameResultPage/components/GameResult/components/index.ts
